Don't zero disk files that don't exist

+++ bench/start-daisy-nfsd.sh
@@ -53,7 +53,9 @@ done
 
 make --quiet compile
 go build ./cmd/daisy-nfsd
-dd if=/dev/zero of="$disk_path" bs=4k count=$((size_mb * 1024 / 4)) 1>/dev/null 2>&1
+if [ -e "$disk_path" ]; then
+    dd if=/dev/zero of="$disk_path" bs=4k count=$((size_mb * 1024 / 4)) 1>/dev/null 2>&1
+fi
 ./daisy-nfsd -debug=0 -disk "$disk_path" -size "$size_mb" "${extra_args[@]}" 1>nfs.out 2>&1 &
 sleep 2
 killall -0 daisy-nfsd       # make sure server is running
